Implement dapp redesign
