+++ src/analytics/Events.tsx
@@ -543,6 +543,7 @@ export enum DappExplorerEvents {
   dapp_unfavorite = 'dapp_unfavorite',
   dapp_filter = 'dapp_filter',
   dapp_rankings_open = 'dapp_rankings_open',
+  dapp_explore_all = 'dapp_explore_all',
 }
 
 export enum WebViewEvents {
